requirements.txt
flask==3.1.1
flask-cors==5.0.0


flask_sqlalchemy


supabase


python-dotenv

